feat: separate app to modules
