+++ README.md
@@ -1,1 +1,3 @@
 # virtual-keyboard
+
+https://github.com/rolling-scopes-school/tasks/blob/master/tasks/codejam-virtual-keyboard.md
